API BREAK: gvc: make 'GVJ_t' a 'size_t'
This is a step towards pervasively using `size_t` for counting. This also
unravels a use of the `ARRAY_SIZE` macro which we should probably move towards
no longer shipping in a public header as it is likely to conflict with user
symbols.

+++ lib/gvc/gvcjob.h
@@ -354,7 +354,7 @@ extern "C" {
 
         /* keybindings for keyboard events */
 	gvevent_key_binding_t *keybindings;
-	int numkeys;
+	size_t numkeys;
 	void *keycodes;
     };
